+++ sdk/python/pulumi_aws/dms/endpoint.py
@@ -49,7 +49,7 @@ class Endpoint(pulumi.CustomResource):
     """
     The password to be used to login to the endpoint database.
     """
-    port: pulumi.Output[int]
+    port: pulumi.Output[float]
     """
     The port used by the endpoint database.
     """
@@ -95,7 +95,7 @@ class Endpoint(pulumi.CustomResource):
         :param pulumi.Input[str] kms_key_arn: The Amazon Resource Name (ARN) for the KMS key that will be used to encrypt the connection parameters. If you do not specify a value for `kms_key_arn`, then AWS DMS will use your default encryption key. AWS KMS creates the default encryption key for your AWS account. Your AWS account has a different default encryption key for each AWS region.
         :param pulumi.Input[dict] mongodb_settings: Settings for the source MongoDB endpoint. Available settings are `auth_type` (default: `PASSWORD`), `auth_mechanism` (default: `DEFAULT`), `nesting_level` (default: `NONE`), `extract_doc_id` (default: `false`), `docs_to_investigate` (default: `1000`) and `auth_source` (default: `admin`). For more details, see [Using MongoDB as a Source for AWS DMS](https://docs.aws.amazon.com/dms/latest/userguide/CHAP_Source.MongoDB.html).
         :param pulumi.Input[str] password: The password to be used to login to the endpoint database.
-        :param pulumi.Input[int] port: The port used by the endpoint database.
+        :param pulumi.Input[float] port: The port used by the endpoint database.
         :param pulumi.Input[dict] s3_settings: Settings for the target S3 endpoint. Available settings are `service_access_role_arn`, `external_table_definition`, `csv_row_delimiter` (default: `\\n`), `csv_delimiter` (default: `,`), `bucket_folder`, `bucket_name` and `compression_type` (default: `NONE`). For more details, see [Using Amazon S3 as a Target for AWS Database Migration Service](https://docs.aws.amazon.com/dms/latest/userguide/CHAP_Target.S3.html).
         :param pulumi.Input[str] server_name: The host name of the server.
         :param pulumi.Input[str] service_access_role: The Amazon Resource Name (ARN) used by the service access IAM role for dynamodb endpoints.
