+++ sdk/python/pulumi_aws/iam/role.py
@@ -29,7 +29,7 @@ class Role(pulumi.CustomResource):
     """
     Specifies to force detaching any policies the role has before destroying it. Defaults to `false`.
     """
-    max_session_duration: pulumi.Output[int]
+    max_session_duration: pulumi.Output[float]
     """
     The maximum session duration (in seconds) that you want to set for the specified role. If you do not specify a value for this setting, the default maximum of one hour is applied. This setting can have a value from 1 hour to 12 hours.
     """
@@ -67,7 +67,7 @@ class Role(pulumi.CustomResource):
         :param pulumi.Input[str] assume_role_policy: The policy that grants an entity permission to assume the role.
         :param pulumi.Input[str] description: The description of the role.
         :param pulumi.Input[bool] force_detach_policies: Specifies to force detaching any policies the role has before destroying it. Defaults to `false`.
-        :param pulumi.Input[int] max_session_duration: The maximum session duration (in seconds) that you want to set for the specified role. If you do not specify a value for this setting, the default maximum of one hour is applied. This setting can have a value from 1 hour to 12 hours.
+        :param pulumi.Input[float] max_session_duration: The maximum session duration (in seconds) that you want to set for the specified role. If you do not specify a value for this setting, the default maximum of one hour is applied. This setting can have a value from 1 hour to 12 hours.
         :param pulumi.Input[str] name: The name of the role. If omitted, Terraform will assign a random, unique name.
         :param pulumi.Input[str] name_prefix: Creates a unique name beginning with the specified prefix. Conflicts with `name`.
         :param pulumi.Input[str] path: The path to the role.
