+++ sdk/python/pulumi_aws/opsworks/memcached_layer.py
@@ -9,7 +9,7 @@ import pulumi.runtime
 from .. import utilities, tables
 
 class MemcachedLayer(pulumi.CustomResource):
-    allocated_memory: pulumi.Output[int]
+    allocated_memory: pulumi.Output[float]
     """
     Amount of memory to allocate for the cache on each instance, in megabytes. Defaults to 512MB.
     """
@@ -58,7 +58,7 @@ class MemcachedLayer(pulumi.CustomResource):
     """
     Whether to install OS and package updates on each instance when it boots.
     """
-    instance_shutdown_timeout: pulumi.Output[int]
+    instance_shutdown_timeout: pulumi.Output[float]
     """
     The time, in seconds, that OpsWorks will wait for Chef to complete after triggering the Shutdown event.
     """
@@ -84,7 +84,7 @@ class MemcachedLayer(pulumi.CustomResource):
         
         :param str resource_name: The name of the resource.
         :param pulumi.ResourceOptions opts: Options for the resource.
-        :param pulumi.Input[int] allocated_memory: Amount of memory to allocate for the cache on each instance, in megabytes. Defaults to 512MB.
+        :param pulumi.Input[float] allocated_memory: Amount of memory to allocate for the cache on each instance, in megabytes. Defaults to 512MB.
         :param pulumi.Input[bool] auto_assign_elastic_ips: Whether to automatically assign an elastic IP address to the layer's instances.
         :param pulumi.Input[bool] auto_assign_public_ips: For stacks belonging to a VPC, whether to automatically assign a public IP address to each of the layer's instances.
         :param pulumi.Input[bool] auto_healing: Whether to enable auto-healing for the layer.
@@ -95,7 +95,7 @@ class MemcachedLayer(pulumi.CustomResource):
         :param pulumi.Input[list] ebs_volumes: `ebs_volume` blocks, as described below, will each create an EBS volume and connect it to the layer's instances.
         :param pulumi.Input[str] elastic_load_balancer: Name of an Elastic Load Balancer to attach to this layer
         :param pulumi.Input[bool] install_updates_on_boot: Whether to install OS and package updates on each instance when it boots.
-        :param pulumi.Input[int] instance_shutdown_timeout: The time, in seconds, that OpsWorks will wait for Chef to complete after triggering the Shutdown event.
+        :param pulumi.Input[float] instance_shutdown_timeout: The time, in seconds, that OpsWorks will wait for Chef to complete after triggering the Shutdown event.
         :param pulumi.Input[str] name: A human-readable name for the layer.
         :param pulumi.Input[str] stack_id: The id of the stack the layer will belong to.
         :param pulumi.Input[list] system_packages: Names of a set of system packages to install on the layer's instances.
